InputView와 OutputView로 역할을 나눔
