+++ CoordinateCalculator/CoordinateCalculator/OutputView.swift
@@ -24,7 +24,7 @@ struct OutputView {
         print(ANSICode.clear + ANSICode.text.cyanBright + ANSICode.CoordinateGrid.draw())
     }
     
-    static func draw(point: Point) throws {
+    private static func draw(point: Point) throws {
        guard  ANSICode.CoordinateGrid.gridRange.contains(point.x),
         ANSICode.CoordinateGrid.gridRange.contains(point.y) else {
             throw OutputError.notDrawableRange
@@ -38,6 +38,35 @@ struct OutputView {
             try draw(point: point)
         }
     }
+    
+    static func shaped(points: [Point]) throws -> Drawable {
+        switch points.count {
+        case 1:
+            return points[0]
+        case 2:
+            return Line(start: points[0], end: points[1])
+        case 3:
+            return Triangle(pointA: points[0], pointB: points[1], pointC: points[2])
+        default:
+            throw OutputError.shapeNotSupported
+        }
+    }
+    
+    static func printShape(_ shape: Drawable) throws {
+        switch shape {
+        case let shape as CustomStringConvertible:
+            try draw(shape: shape as! Drawable)
+            printText(shape.description)
+        default:
+            try draw(shape: shape)
+        }
+    }
+    
+    
+    
+    static func printText(_ text: String) {
+        print(ANSICode.cursor.move(row: ANSICode.CoordinateGrid.gridLimit + 5, column: 1) + text)
+    }
 
     
 }
@@ -45,11 +74,15 @@ struct OutputView {
 enum OutputError: Error, CustomStringConvertible {
     
     case notDrawableRange
+    case shapeNotSupported
     
     var description: String {
         switch self {
         case .notDrawableRange:
             return "그릴 수 있는 범위가 아닙니다."
+        case .shapeNotSupported:
+            return "지원되지 않는 도형입니다."
         }
     }
+    
 }
