+++ CoordinateCalculator/CoordinateCalculator/main.swift
@@ -3,20 +3,23 @@ import Foundation
 func main() {
     
     OutputView.drawCoordinatePlane()
-    var textView = InputView()
-    var coordinateCalculator = CoordinateCalCulator()
+    
     
     do {
         
+        let points = try InputView.readCoordinates()
+        let shape = try OutputView.shaped(points: points)
+        try OutputView.printShape(shape)
+        
     } catch let error as InputError {
-        textView.printText("입력 오류: \(error.description)")
-    } catch let error as CoordinateCalculatorError {
-        textView.printText("좌표 계산기 오류: \(error.description)")
+        OutputView.printText("입력 오류: \(error.description)")
+    } catch let error as OutputError {
+        OutputView.printText("출력 오류: \(error.description)")
     } catch {
-        textView.printText("예상치 못한 오류: \(error)")
+        OutputView.printText("예상치 못한 오류: \(error)")
     }
     
-    
+    print(ANSICode.cursor.move(row: ANSICode.CoordinateGrid.gridLimit + 5, column: 1))
     
 }
